add ICS download for personal events
