+++ Sources/swiftarr/Site/SiteEventsController.swift
@@ -92,6 +92,7 @@ struct SiteEventsController: SiteControllerUtils {
 		let privateRoutes = getPrivateRoutes(app, feature: .schedule)
 		privateRoutes.post("events", eventIDParam, "favorite", use: eventsAddRemoveFavoriteHandler)
 		privateRoutes.delete("events", eventIDParam, "favorite", use: eventsAddRemoveFavoriteHandler)
+		privateRoutes.get("events", "personal", personalEventIDParam, "calendarevent.ics", use: personalEventDownloadICSHandler)
 	}
 
 	// MARK: - Events
@@ -195,7 +196,23 @@ struct SiteEventsController: SiteControllerUtils {
 		let response = try await apiQuery(req, endpoint: "/events/\(eventID)")
 		let event = try response.content.decode(EventData.self)
 		let username = req.auth.get(UserCacheData.self)?.username ?? ""
-		let icsString = buildEventICS(events: [event], username: username)
+		let icsString = ICSHelper.buildEventICS(events: [event], username: username)
+		let cleanEventTitle = event.title.replacingOccurrences(of: "\"", with: "")
+		let headers = HTTPHeaders([("Content-Disposition", "attachment; filename=\"\(cleanEventTitle).ics\"")])
+		return try await icsString.encodeResponse(status: .ok, headers: headers, for: req)
+	}
+
+	// `GET /events/personal/:personal_event_id/calendarevent.ics`
+	//
+	// Returns a .ics file containing info on the given event; suitable for opening in calendaring apps.
+	func personalEventDownloadICSHandler(_ req: Request) async throws -> Response {
+		guard let eventID = req.parameters.get(personalEventIDParam.paramString)?.percentEncodeFilePathEntry() else {
+			throw Abort(.badRequest, reason: "Missing event ID parameter.")
+		}
+		let response = try await apiQuery(req, endpoint: "/personalevents/\(eventID)")
+		let event = try response.content.decode(PersonalEventData.self)
+		let username = req.auth.get(UserCacheData.self)?.username ?? ""
+		let icsString = ICSHelper.buildPersonalEventICS(events: [event], username: username)
 		let cleanEventTitle = event.title.replacingOccurrences(of: "\"", with: "")
 		let headers = HTTPHeaders([("Content-Disposition", "attachment; filename=\"\(cleanEventTitle).ics\"")])
 		return try await icsString.encodeResponse(status: .ok, headers: headers, for: req)
@@ -232,7 +249,7 @@ struct SiteEventsController: SiteControllerUtils {
 
 		let response = try await apiQuery(req, endpoint: "/events/favorites")
 		let events = try response.content.decode([EventData].self)
-		let icsString = buildEventICS(events: events, username: user.username)
+		let icsString = ICSHelper.buildEventICS(events: events, username: user.username)
 		let cleanEventTitle = "JoCo Cruise: \(user.username)"
 		let headers = HTTPHeaders([
 			("Content-Disposition", "attachment; filename=\"\(cleanEventTitle).ics\""),
@@ -248,67 +265,6 @@ struct SiteEventsController: SiteControllerUtils {
 		}
 		let response = try await apiQuery(req, endpoint: "/events/\(eventID)/favorite", method: req.method)
 		return response.status
-	}
-
-	// MARK: - Utility fns
-
-	/// Creates a iCalendar data file describing the given event. iCalendar is also known as VCALENDAR or an .ics file.
-	/// It's the thing most calendar event importers have standardized on for data interchange.
-	func buildEventICS(events: [EventData], username: String = "") -> String {
-		let yearFormatter = DateFormatter()
-		yearFormatter.setLocalizedDateFormatFromTemplate("y")
-		let cruiseYear = yearFormatter.string(from: Settings.shared.cruiseStartDate())
-		var resultICSString = """
-										BEGIN:VCALENDAR
-										VERSION:2.0
-										X-WR-CALNAME:JoCo Cruise \(cruiseYear): \(username)
-										X-WR-CALDESC:Event Calendar
-										METHOD:PUBLISH
-										CALSCALE:GREGORIAN
-										PRODID:-//Sched.com JoCo Cruise \(cruiseYear)//EN
-										X-WR-TIMEZONE:UTC
-										
-										"""
-		let dateFormatter = ISO8601DateFormatter()
-		dateFormatter.formatOptions = [.withYear, .withMonth, .withDay, .withTime, .withTimeZone]
-		for event in events {
-			let startTime = dateFormatter.string(from: event.startTime)
-			let endTime = dateFormatter.string(from: event.endTime)
-			let stampTime = dateFormatter.string(from: event.lastUpdateTime)  // DTSTAMP is when the ICS was last modified.
-			resultICSString.append(
-				"""
-													BEGIN:VEVENT
-													DTSTAMP:\(stampTime)
-													DTSTART:\(startTime)
-													DTEND:\(endTime)
-													SUMMARY:\(icsEscapeString(event.title))
-													DESCRIPTION:\(icsEscapeString(event.description))
-													CATEGORIES:\(icsEscapeString(event.eventType))
-													LOCATION:\(icsEscapeString(event.location))
-													SEQUENCE:0
-													UID:\(icsEscapeString(event.uid))
-													URL:https://jococruise\(cruiseYear).sched.com/event/\(event.uid)
-													END:VEVENT
-													
-													"""
-			)
-		}
-		resultICSString.append(
-			"""
-										END:VCALENDAR
-										
-										"""
-		)
-		return resultICSString
-	}
-
-	// the ICS file format has specific string escaping requirements. See https://datatracker.ietf.org/doc/html/rfc5545
-	func icsEscapeString(_ str: String) -> String {
-		let result = str.replacingOccurrences(of: "\\", with: "\\\\")
-			.replacingOccurrences(of: ";", with: "\\;")
-			.replacingOccurrences(of: ",", with: "\\,")
-			.replacingOccurrences(of: "\n", with: "\\n")
-		return result
 	}
 }
